+++ cogs/games/example.py
@@ -1,6 +1,6 @@
 """Sub-cog example."""
 import gearbox
-cog = gearbox.Cog('example')
+cog = gearbox.Cog()
 
 
 @cog.command