+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 + Displaying version number in *Playing* status
 + Cog-specific configuration files (yaml or json)
 
+#### Fixed
+
+* Not naming a sub-cog explicitly may crash the bot
+
 ### [[v0.1.3](https://github.com/Zeroji/semicolon/releases/tag/v0.1.3)]
 
 #### Added